Fixes crossplatform issues.
