+++ lib/motion-i18n.rb
@@ -1,12 +1,27 @@
 require 'motion-require'
-require 'tasks/translate'
 
 platform = Motion::Project::App.respond_to?(:template) ? Motion::Project::App.template : :ios
 if platform.to_s.start_with?('ios') or platform.to_s.start_with?('osx')
-  platform = 'ios'
+  platform_name = 'ios'
 elsif platform.to_s.start_with?('android')
-  platform = 'droid'
+  platform_name = 'android'
 end
 
-Motion::Require.all(Dir.glob(File.expand_path("../motion-i18n/**/**_#{platform}.rb", __FILE__)))
-Motion::Require.all(Dir.glob(File.expand_path('../motion-i18n/**/**.rb', __FILE__)))
+require "motion-i18n-#{platform_name}/translate"
+
+Motion::Project::App.setup do |app|
+  platform_lib = File.join(File.dirname(__FILE__), "motion-i18n-#{platform_name}")
+  unless File.exists? platform_lib
+    raise "Sorry, the platform #{platform.inspect} (aka #{platform_name}) is not supported by motion-i18n"
+  end
+
+  # scans app.files until it finds app/ (the default)
+  # if found, it inserts just before those files, otherwise it will insert to
+  # the end of the list
+  insert_point = app.files.find_index { |file| file =~ /^(?:\.\/)?app\// } || 0
+
+  Dir.glob(File.join(platform_lib, 'i18n.rb')).reverse.each do |file|
+    app.files.insert(insert_point, file)
+  end
+
+end
